TDI-41630 : fix reject handling with snowflake copy

+++ jdbc/src/main/java/org/talend/components/jdbc/output/statement/operations/snowflake/SnowflakeDelete.java
@@ -23,12 +23,9 @@ import org.talend.sdk.component.api.record.Record;
 import java.sql.Connection;
 import java.sql.SQLException;
 import java.sql.Statement;
-import java.time.ZoneOffset;
-import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
 import java.util.List;
 
-import static java.time.LocalDateTime.now;
 import static java.util.stream.Collectors.joining;
 import static org.talend.components.jdbc.output.statement.operations.snowflake.SnowflakeCopy.putAndCopy;
 import static org.talend.components.jdbc.output.statement.operations.snowflake.SnowflakeCopy.tmpTableName;
@@ -51,11 +48,12 @@ public class SnowflakeDelete extends Delete {
             final String fqTmpTableName = namespace(connection) + "." + getPlatform().identifier(tmpTableName);
             final String fqStageName = namespace(connection) + ".%" + getPlatform().identifier(tmpTableName);
             rejects.addAll(putAndCopy(connection, records, fqStageName, fqTableName, fqTmpTableName));
-            // delete from temp table
-            try (final Statement statement = connection.createStatement()) {
-                statement.execute("delete from " + fqTableName + " target using " + fqTmpTableName + " as source where "
-                        + getConfiguration().getKeys().stream().map(key -> getPlatform().identifier(key))
-                                .map(key -> "source." + key + "= target." + key).collect(joining("AND", " ", " ")));
+            if (records.size() != rejects.size()) {
+                try (final Statement statement = connection.createStatement()) {
+                    statement.execute("delete from " + fqTableName + " target using " + fqTmpTableName + " as source where "
+                            + getConfiguration().getKeys().stream().map(key -> getPlatform().identifier(key))
+                                    .map(key -> "source." + key + "= target." + key).collect(joining("AND", " ", " ")));
+                }
             }
             connection.commit();
         }
